Create hierarchical structure for Nodes
