+++ src/main/java/GUI/GraphDrawer.java
@@ -1,7 +1,7 @@
 package GUI;
 
 import graph.SequenceGraph;
-import parser.SequenceNode;
+import graph.SequenceNode;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Color;
