+++ src/main/java/graph/DummyNode.java
@@ -1,0 +1,11 @@
+package graph;
+
+/**
+ * Created by TUDelft SID on 8-5-2017.
+ */
+public class DummyNode extends Node {
+
+    public DummyNode(int id) {
+        super(id);
+    }
+}
