+++ src/main/java/graph/Edge.java
@@ -1,4 +1,4 @@
-package parser;
+package graph;
 
 import java.util.ArrayList;
