+++ src/main/java/parser/GfaParser.java
@@ -1,6 +1,8 @@
 package parser;
 
+import graph.Edge;
 import graph.SequenceGraph;
+import graph.SequenceNode;
 
 import java.io.*;
 import java.util.ArrayList;
